Just organize components involved with drawing on the map
